Restructuring for better modularity
